Initial expression concrete classes

+++ app/src/main/java/com/example/msharp/expression/Expression.java
@@ -6,7 +6,7 @@ public abstract class Expression extends AST {
     public abstract void execute() throws Exception;
     public abstract int type() throws Exception;
     public abstract int resultInt();
-    public abstract int resultDouble();
+    public abstract double resultDouble();
     public abstract String resultBool();
     public abstract String resultString();
 }
